// src/app/models/aviso-mp.model.ts
export class AvisoMP {
	// id?: number;
	// fecElaboracion?: string;
	// desDelegacionMunicipio?: string;
	// desAgenciaMp?: string;
	// nomPaciente?: string;
	// desUnidadMedica?: string;
	// ubicacionHospital?: string;
	// cveEspecialidad?: string;
	// cveCama?: string;
	// fecIngreso?: string;
	// timIngreso?: string;
	// desLesionesPaciente?: string;
	// nomMedico?: string;
	// desMatriculaMedico?: string;
	// nomTrabajadorSocial?: string;
	// desMatriculaTs?: string;
	// idEstado?: string;

	fechaRegistroAviso?: string;
	idEstado?: string;
	alcaldia?: string;
	agenciaMP?: string;
	nombrePaciente?: string;
	idUnidadHospital?: string;
	ubicacionHospital?: string;
	cveServicio?: string;
	numCama?: string;
	fechaIngreso?: string;
	horaIngreso?: string;
	lesionesPaciente?: string;
	nombreMedicoTratante?: string;
	matriculaMedicoTratante?: string;
	nombreTrabajadorSocial?: string;
	matriculaTrabajadorSocial?: string;
}